[FEAT] BOM 코드, 총비용, 구성품 수, 버전 추가

+++ src/main/java/com/sampoom/backend/api/item/dto/ItemResponseDTO.java
@@ -27,7 +27,7 @@ public class ItemResponseDTO {
     private Integer leadTime;
     private Integer baseQuantity;
 
-    private BigDecimal standardCost;
+    private Long standardCost;
 
     public static ItemResponseDTO ofMaterial(MaterialResponseDTO m) {
         return ItemResponseDTO.builder()
