+++ src/main/java/com/sampoom/backend/api/material/dto/MaterialResponseDTO.java
@@ -18,7 +18,7 @@ public class MaterialResponseDTO {
     private Integer baseQuantity;  // 기준단위
     private Integer leadTime;      // 리드타임
 
-    private BigDecimal standardCost;         // 표준단가
+    private Long standardCost;         // 표준단가
 
     private Long materialCategoryId;
     private String materialCategoryName;
