+++ src/main/java/com/sampoom/backend/api/part/dto/PartResponseDTO.java
@@ -16,7 +16,7 @@ public class PartResponseDTO {
     private Integer baseQuantity;
     private Integer leadTime;
 
-    private BigDecimal standardCost;
+    private Long standardCost;
 
     private Long groupId;
     private String groupName;
